==> reference/04_RDD_Shared Variables.ipynb <==
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

# Shared Variables

- 모든 노드에서 사용하기 위한 공유변수


- 공유변수로 지정한 값은 모든 노드에 중복되어 캐시된다.


- 반복적으로 사용해야하는 변수라면,  
  스파크의 노드는 네트워크를 통해 통신 하기 때문에 모든 노드에 중복 캐시하는 시스템적 비용보다  
  네트워크 과정에서 발생하는 오버헤드 비용이 더 많이 발생하게 된다.

## Broadcast Variables

- 각 노드에 공유되는 읽기 전용 변수

#학생 별 수업 카테고리 코드로 지정되어 있는 값을 카테고리의 상세값으로 변경한다고 가정
data = [("홍길동","DE"),\
    ("이제동","DS"),\
    ("하명도","DE"),\
    ("변현재","WD")]
code_desc = {"DE":"Data Engineer", "DS":"Data Science", "WD":"Web Developer"}
#sparkContext.broadcast()
#Key값으로 조인해서 반환-파이썬
students_rdd=sc.parallelize(data)
students_rdd.mapValues(lambda e : code_desc[e]).collect()

#Broadcast_variables 사용하기
broadcast_code = sc.broadcast(code_desc)
broadcast_code.value['DE'] #원하는값 찍어서 나옴
#공유변수에서 DE 키값을 삭제하더라도 코드변환이 이루어진다. 즉 삭제가 되지 않음
broadcast_code.value
del(broadcast_code.value['DE']) #변수자체를 지우면 주소가 지워져서 없엉지나, 값을 지정해서 삭제하면 없어지지 않음
#broadcast 함수를 사용해 생성하는 시점에 sparkContext에 등록되기 때문
students_rdd.mapValues(lambda e : broadcast_code.value[e]).collect()
#스파크 세션 거쳐서


## Accumulator

- 각 노드에 공유되는 누산기 함수

accum = sc.accumulator(0) #공유변수,전역변수
def change_cate(e):
    accum.add(1)
    return broadcast_code.value[e]

students_rdd.mapValues(lambda e : change_cate(e)).collect()
accum.value #4바퀴 돌았음
